Revamp decode error handling; NOT backwards compatible (#50)
The new decode error handling uses an internal error state so calling code to decode a token is simpler.

All the claim decoding functions now return void. You can adapt old code to this by adding a get error call after the claim decode.

There's a few other fixes here too.

This is not a backwards compatible change, but adapting isn't too hard. This is hopefully the last change to ctoken that not backwards compatible.


* new error handling for decode starting to work

* more conversion to new error handling

* almost all converted to new error scheme

* error handling for submods

* all working for new errors; documentation NOT updated

* starting work on documentation

* still working on documentation

* almost done

* completed TODO's related to decode errors; it might be done

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>